Fix autotools building
Previously this project was only built with build2.

Perhaps it's an artifact of how build2 works: every #include used angled
brackets, such as:

This is incorrect, because it instructs the preprocessort to *search* all
include paths if a specific file is missing.

The correct behaviour is for the preprocessor to look for files owned by
this project in a single location, and fail if that file does not exist.
There should be no searching.

The problem with unnecessary searching is:
1. a file with the right name might be found, though it is not the intended
   include. Depending on its contents, this could cause subtle bugs, eg,
   if it contains matching definitions, but not the wanted definition.
   (ie, this is a correctness problem)
2. Searching for files through the potentially multiple search paths,
   rather than opening the owned file directly, causes unnecessary disk IO
   thus slowing down not only the compiler, but also other simultaneous
   compile jobs competing for the same I/O
   (ie, this is a performance problem)

This commit replaces all angled-brackets includes quoted includes. The
autotools scripts never included the srcdir with -I<dir>, which is how
this problem became obvious.

This commit also adds libcutl as submodule, as it is a dependency.